Fix the util_tests.cpp build on OSX
Summary: The osx compiler is confused with the Amount multiplication operator. Should it be int or int64_t ?  Let's hint the compiler with a good'ol cast.

Test Plan:
On OSX:
  ninja check

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11362

+++ src/test/util_tests.cpp
@@ -1500,7 +1500,7 @@ BOOST_AUTO_TEST_CASE(util_ParseMoney) {
 
     // Max possible value with the max number of digits
     BOOST_CHECK(ParseMoney("9999999999999999.99", ret));
-    BOOST_CHECK_EQUAL(ret, 999999999999999999 * SATOSHI);
+    BOOST_CHECK_EQUAL(ret, int64_t(999999999999999999) * SATOSHI);
     // Attempted 63 bit overflow should fail
     BOOST_CHECK(!ParseMoney("10000000000000000.00", ret));
     BOOST_CHECK(!ParseMoney("92233720368547758.08", ret));
